UI and core changes to support the current/voltage correction.
Since the correction parameters are stored into the controller's flash,
we do not want the same behaviour as for the max_xxx parameters, which
are always send on startup from pypilot to the controller.

Instead, we want to retrieve first the stored values, and only then
allow modifications from the UI which needs to be send back to the
controller.