Fix key configuration page.
Recent changes broke the fact that the config['actions'] map must be
kept ordered, because the web interface depends on that.

Also, the space in the name of pilot modes and profiles is confusing
javascript.

Now the web page for key configuration is listing all the items in the
right order, and IR events can be assigned to each of them.